Fixed remove bug where node has 2 children.
When removing a node that has 2 children, and it's left subtree has no right descendants, the removal of the replacement from it's current position needs to be set on the replacement parents left pointer. This is due to the fact that it was found on the left pointer. Compared to the more normal condition where the left sub tree's with right children, that relationship will be found on the right pointer.